feat: Add Facebook Test Event Code setting
Introduces a new setting for the Facebook Test Event Code to facilitate event testing in Facebook Events Manager. This change includes updating the AppSettings type, the default settings in the store, and integrating the new code into the Facebook Pixel injection logic within the App component. The Dashboard page also receives an update to include an input field for this new setting.